Changed engine/Fixed a potential bug in main file
There was some unncessary code that I had previously placed in `engine.c`. There was also some code that I realized wasn't working the way that I had anticipated, so I fixed it.

I added an `if` statement to `gtktactoe.c` that will prevent a potential bug when starting a new game before the current game is finished.